Add more generic ban handling
