+++ Cargo.toml
@@ -28,6 +28,7 @@ notify-debouncer-full = "0.3.1"
 reborrow = { version="0.5.5", features=["derive"] }
 anyhow = "1"
 arraydeque = "0.5"
+cached = "0.51"
 
 [build-dependencies]
 vergen-git2 = { version = "1.0.0-beta.2", features = [] }
